+++ nullaway/src/main/java/com/uber/nullaway/handlers/StreamNullabilityPropagator.java
@@ -24,11 +24,7 @@ package com.uber.nullaway.handlers;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
 import com.google.errorprone.VisitorState;
-import com.google.errorprone.predicates.TypePredicate;
-import com.google.errorprone.predicates.type.DescendantOf;
-import com.google.errorprone.suppliers.Suppliers;
 import com.google.errorprone.util.ASTHelpers;
 import com.sun.source.tree.ClassTree;
 import com.sun.source.tree.ExpressionTree;
@@ -52,7 +48,9 @@ import com.uber.nullaway.dataflow.AccessPath;
 import com.uber.nullaway.dataflow.AccessPathElement;
 import com.uber.nullaway.dataflow.AccessPathNullnessAnalysis;
 import com.uber.nullaway.dataflow.NullnessStore;
-import java.util.ArrayList;
+import com.uber.nullaway.handlers.stream.MaplikeMethodRecord;
+import com.uber.nullaway.handlers.stream.MaplikeToFilterInstanceRecord;
+import com.uber.nullaway.handlers.stream.StreamTypeRecord;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -72,33 +70,6 @@ import org.checkerframework.dataflow.cfg.node.LocalVariableNode;
  * observable.filter(... return o.foo() != null; ...).map(... o.foo.toString() ...)
  */
 class StreamNullabilityPropagator extends BaseNoOpHandler {
-
-  private final ImmutableList<StreamTypeRecord> STREAM_MODELS =
-      StreamModelBuilder.start()
-          .addStreamType(new DescendantOf(Suppliers.typeFromString("java.util.stream.Stream")))
-          // Names of all the methods of java.util.stream.Stream that behave like .filter(...)
-          // (must take exactly 1 argument)
-          .withFilterMethodFromSignature("filter(java.util.function.Predicate<? super T>)")
-          // Names and relevant arguments of all the methods of java.util.stream.Stream that behave
-          // like .map(...) for the purposes of this checker (the listed arguments are those that
-          // take the potentially filtered objects from the stream)
-          .withMapMethodFromSignature(
-              "<R>map(java.util.function.Function<? super T,? extends R>)",
-              "apply",
-              ImmutableSet.of(0))
-          .withMapMethodAllFromName("flatMap", "apply", ImmutableSet.of(0))
-          // List of methods of java.util.stream.Stream through which we just propagate the
-          // nullability information of the last call, e.g. m() in
-          // Observable.filter(...).m().map(...) means the
-          // nullability information from filter(...) should still be propagated to map(...),
-          // ignoring the interleaving call to m().
-          .withPassthroughMethodFromSignature("distinct()")
-          // List of methods of java.util.stream.Stream that both use the nullability information
-          // internally (like map does), but also don't change the values flowing through the stream
-          // and thus propagate
-          // the nullability information of the last call.
-
-          .end();
 
   /* Terminology for this class internals:
    *
@@ -129,23 +100,22 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
    */
 
   // Set of filter methods found thus far (e.g. A.filter, see above)
-  private final Set<Tree> filterMethodOrLambdaSet = new LinkedHashSet<Tree>();
+  private final Set<Tree> filterMethodOrLambdaSet = new LinkedHashSet<>();
 
   // Maps each call in the observable call chain to its outer call (see above).
   private final Map<MethodInvocationTree, MethodInvocationTree> observableOuterCallInChain =
-      new LinkedHashMap<MethodInvocationTree, MethodInvocationTree>();
+      new LinkedHashMap<>();
 
   // Maps the call in the observable call chain to the relevant inner method or lambda.
   // e.g. In the example above:
   //   observable.filter() => A.filter
   //   observable.filter().map() => B.apply
   private final Map<MethodInvocationTree, Tree> observableCallToInnerMethodOrLambda =
-      new LinkedHashMap<MethodInvocationTree, Tree>();
+      new LinkedHashMap<>();
 
   // Map from map method (or lambda) to corresponding previous filter method (e.g. B.apply =>
   // A.filter)
-  private final Map<Tree, MaplikeToFilterInstanceRecord> mapToFilterMap =
-      new LinkedHashMap<Tree, MaplikeToFilterInstanceRecord>();
+  private final Map<Tree, MaplikeToFilterInstanceRecord> mapToFilterMap = new LinkedHashMap<>();
 
   /*
    * Note that the above methods imply a diagram like the following:
@@ -165,24 +135,25 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
   // true.
   // Specifically, this is the least upper bound of the "then" store on the branch of every return
   // statement in which the expression after the return can be true.
-  private final Map<Tree, NullnessStore> filterToNSMap = new LinkedHashMap<Tree, NullnessStore>();
+  private final Map<Tree, NullnessStore> filterToNSMap = new LinkedHashMap<>();
 
   // Maps the body of a method or lambda to the corresponding enclosing tree, used because the
   // dataflow analysis
   // loses the pointer to the tree by the time we hook into its body.
-  private final Map<Tree, Tree> bodyToMethodOrLambda = new LinkedHashMap<Tree, Tree>();
+  private final Map<Tree, Tree> bodyToMethodOrLambda = new LinkedHashMap<>();
 
   // Maps the return statements of the filter method to the filter tree itself, similar issue as
   // above.
-  private final Map<ReturnTree, Tree> returnToEnclosingMethodOrLambda =
-      new LinkedHashMap<ReturnTree, Tree>();
+  private final Map<ReturnTree, Tree> returnToEnclosingMethodOrLambda = new LinkedHashMap<>();
 
   // Similar to above, but mapping expression-bodies to their enclosing lambdas
   private final Map<ExpressionTree, LambdaExpressionTree> expressionBodyToFilterLambda =
-      new LinkedHashMap<ExpressionTree, LambdaExpressionTree>();
+      new LinkedHashMap<>();
+  private final ImmutableList<StreamTypeRecord> models;
 
-  StreamNullabilityPropagator() {
+  StreamNullabilityPropagator(ImmutableList<StreamTypeRecord> models) {
     super();
+    this.models = models;
   }
 
   @Override
@@ -205,7 +176,7 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
       VisitorState state,
       Symbol.MethodSymbol methodSymbol) {
     Type receiverType = ASTHelpers.getReceiverType(tree);
-    for (StreamTypeRecord streamType : STREAM_MODELS) {
+    for (StreamTypeRecord streamType : models) {
       if (streamType.matchesType(receiverType, state)) {
         // Build observable call chain
         buildObservableCallChain(tree);
@@ -237,9 +208,9 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
               handleMapAnonClass(methodRecord, tree, annonClassBody);
             }
           } else if (argTree instanceof LambdaExpressionTree) {
-            observableCallToInnerMethodOrLambda.put(tree, (LambdaExpressionTree) argTree);
+            observableCallToInnerMethodOrLambda.put(tree, argTree);
           } else if (argTree instanceof MemberReferenceTree) {
-            observableCallToInnerMethodOrLambda.put(tree, (MemberReferenceTree) argTree);
+            observableCallToInnerMethodOrLambda.put(tree, argTree);
           }
         }
       }
@@ -293,7 +264,7 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
     for (Tree t : annonClassBody.getMembers()) {
       if (t instanceof MethodTree && ((MethodTree) t).getName().toString().equals("test")) {
         filterMethodOrLambdaSet.add(t);
-        observableCallToInnerMethodOrLambda.put(observableDotFilter, (MethodTree) t);
+        observableCallToInnerMethodOrLambda.put(observableDotFilter, t);
         handleChainFromFilter(streamType, observableDotFilter, t, state);
       }
     }
@@ -316,7 +287,7 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
     for (Tree t : annonClassBody.getMembers()) {
       if (t instanceof MethodTree
           && ((MethodTree) t).getName().toString().equals(methodRecord.getInnerMethodName())) {
-        observableCallToInnerMethodOrLambda.put(observableDotMap, (MethodTree) t);
+        observableCallToInnerMethodOrLambda.put(observableDotMap, t);
       }
     }
   }
@@ -448,8 +419,8 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
       assert (filterTree instanceof MethodTree || filterTree instanceof LambdaExpressionTree);
       MaplikeMethodRecord mapMR = callInstanceRecord.getMaplikeMethodRecord();
       for (int argIdx : mapMR.getArgsFromStream()) {
-        LocalVariableNode filterLocalName = null;
-        LocalVariableNode mapLocalName = null;
+        LocalVariableNode filterLocalName;
+        LocalVariableNode mapLocalName;
         if (filterTree instanceof MethodTree) {
           filterLocalName = new LocalVariableNode(((MethodTree) filterTree).getParameters().get(0));
         } else {
@@ -499,324 +470,6 @@ class StreamNullabilityPropagator extends BaseNoOpHandler {
       if (canBooleanExpressionEvalToTrue(tree)) {
         filterToNSMap.put(filterTree, thenStore);
       }
-    }
-  }
-
-  /**
-   * Used to produce a new list of StreamTypeRecord models, where each model represents a class from
-   * a stream-based API such as RxJava.
-   *
-   * <p>This class should be used as:
-   *
-   * <p>[...] models = StreamModelBuilder.start() // Start the builder .addStreamType(...) // Add a
-   * type filter matching a stream type .withX(...) // Model the type methods ... .end();
-   */
-  private static class StreamModelBuilder {
-
-    private final List<StreamTypeRecord> typeRecords = new ArrayList<StreamTypeRecord>();
-    private TypePredicate tp = null;
-    private ImmutableSet.Builder<String> filterMethodSigs;
-    private ImmutableSet.Builder<String> filterMethodSimpleNames;
-    private ImmutableMap.Builder<String, MaplikeMethodRecord> mapMethodSigToRecord;
-    private ImmutableMap.Builder<String, MaplikeMethodRecord> mapMethodSimpleNameToRecord;
-    private ImmutableSet.Builder<String> passthroughMethodSigs;
-    private ImmutableSet.Builder<String> passthroughMethodSimpleNames;
-
-    private StreamModelBuilder() {}
-
-    /**
-     * Get an empty StreamModelBuilder.
-     *
-     * @return An empty StreamModelBuilder.
-     */
-    public static StreamModelBuilder start() {
-      return new StreamModelBuilder();
-    }
-
-    private void finalizeOpenStreamTypeRecord() {
-      if (this.tp != null) {
-        typeRecords.add(
-            new StreamTypeRecord(
-                this.tp,
-                filterMethodSigs.build(),
-                filterMethodSimpleNames.build(),
-                mapMethodSigToRecord.build(),
-                mapMethodSimpleNameToRecord.build(),
-                passthroughMethodSigs.build(),
-                passthroughMethodSimpleNames.build()));
-      }
-    }
-
-    /**
-     * Add a stream type to our models.
-     *
-     * @param tp A type predicate matching the class/interface of the type in our stream-based API.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder addStreamType(TypePredicate tp) {
-      finalizeOpenStreamTypeRecord();
-      this.tp = tp;
-      this.filterMethodSigs = ImmutableSet.builder();
-      this.filterMethodSimpleNames = ImmutableSet.builder();
-      this.mapMethodSigToRecord = ImmutableMap.builder();
-      this.mapMethodSimpleNameToRecord = ImmutableMap.builder();
-      this.passthroughMethodSigs = ImmutableSet.builder();
-      this.passthroughMethodSimpleNames = ImmutableSet.builder();
-      return this;
-    }
-
-    /**
-     * Add a filter method to the last added stream type.
-     *
-     * @param filterMethodSig The full sub-signature (everything except the receiver type) of the
-     *     filter method.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withFilterMethodFromSignature(String filterMethodSig) {
-      this.filterMethodSigs.add(filterMethodSig);
-      return this;
-    }
-
-    /**
-     * Add all methods of the last stream type with the given simple name as filter methods.
-     *
-     * @param methodSimpleName The method's simple name.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withFilterMethodAllFromName(String methodSimpleName) {
-      this.filterMethodSimpleNames.add(methodSimpleName);
-      return this;
-    }
-
-    /**
-     * Add a model for a map method to the last added stream type.
-     *
-     * @param methodSig The full sub-signature (everything except the receiver type) of the method.
-     * @param innerMethodName The name of the inner "apply" method of the callback or functional
-     *     interface that must be passed to this method.
-     * @param argsFromStream The indexes (starting at 0, not counting the receiver) of all the
-     *     arguments to this method that receive objects from the stream.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withMapMethodFromSignature(
-        String methodSig, String innerMethodName, ImmutableSet<Integer> argsFromStream) {
-      this.mapMethodSigToRecord.put(
-          methodSig, new MaplikeMethodRecord(innerMethodName, argsFromStream));
-      return this;
-    }
-
-    /**
-     * Add all methods of the last stream type with the given simple name as map methods.
-     *
-     * @param methodSimpleName The method's simple name.
-     * @param innerMethodName The name of the inner "apply" method of the callback or functional
-     *     interface that must be passed to this method.
-     * @param argsFromStream The indexes (starting at 0, not counting the receiver) of all the
-     *     arguments to this method that receive objects from the stream. Must be the same for all
-     *     methods with this name (else use withMapMethodFromSignature).
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withMapMethodAllFromName(
-        String methodSimpleName, String innerMethodName, ImmutableSet<Integer> argsFromStream) {
-      this.mapMethodSimpleNameToRecord.put(
-          methodSimpleName, new MaplikeMethodRecord(innerMethodName, argsFromStream));
-      return this;
-    }
-
-    /**
-     * Add a passthrough method to the last added stream type.
-     *
-     * <p>A passthrough method is a method that affects the stream but doesn't change the
-     * nullability information of the elements inside the stream (e.g. in
-     * o.filter(...).sync().map(...), sync() is a passthrough method if the exact same objects that
-     * are added to the stream at the end of filter are those that will be consumed by map(...).
-     *
-     * @param passthroughMethodSig The full sub-signature (everything except the receiver type) of
-     *     the method.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withPassthroughMethodFromSignature(String passthroughMethodSig) {
-      this.passthroughMethodSigs.add(passthroughMethodSig);
-      return this;
-    }
-
-    /**
-     * Add all methods of the last stream type with the given simple name as passthrough methods.
-     *
-     * @param methodSimpleName The method's simple name.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withPassthroughMethodAllFromName(String methodSimpleName) {
-      this.passthroughMethodSimpleNames.add(methodSimpleName);
-      return this;
-    }
-
-    /**
-     * Add a passthrough method that uses the value to the last added stream type.
-     *
-     * <p>Like a normal passthrough method, but it takes a callback which inspects but doesn't
-     * change the elements flowing through the stream.
-     *
-     * @param passthroughMethodSig The full sub-signature (everything except the receiver type) of
-     *     the method.
-     * @param innerMethodName The name of the inner method of the callback or functional interface
-     *     that must be passed to this method.
-     * @param argsFromStream The indexes (starting at 0, not counting the receiver) of all the
-     *     arguments to this method that receive objects from the stream.
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withUseAndPassthroughMethodFromSignature(
-        String passthroughMethodSig, String innerMethodName, ImmutableSet<Integer> argsFromStream) {
-      this.mapMethodSigToRecord.put(
-          passthroughMethodSig, new MaplikeMethodRecord(innerMethodName, argsFromStream));
-      this.passthroughMethodSigs.add(passthroughMethodSig);
-      return this;
-    }
-
-    /**
-     * Add all methods of the last stream type with the given simple name as use-and-passthrough
-     * methods.
-     *
-     * @param methodSimpleName The method's simple name.
-     * @param innerMethodName The name of the inner method of the callback or functional interface
-     *     that must be passed to this method.
-     * @param argsFromStream The indexes (starting at 0, not counting the receiver) of all the
-     *     arguments to this method that receive objects from the stream. Must be the same for all
-     *     methods with this name (else use withUseAndPassthroughMethodFromSignature).
-     * @return This builder (for chaining).
-     */
-    public StreamModelBuilder withUseAndPassthroughMethodAllFromName(
-        String methodSimpleName, String innerMethodName, ImmutableSet<Integer> argsFromStream) {
-      this.mapMethodSimpleNameToRecord.put(
-          methodSimpleName, new MaplikeMethodRecord(innerMethodName, argsFromStream));
-      this.passthroughMethodSimpleNames.add(methodSimpleName);
-      return this;
-    }
-
-    /**
-     * Turn the models added to this builder into a list of StreamTypeRecord objects.
-     *
-     * @return The finalized (immutable) models.
-     */
-    public ImmutableList<StreamTypeRecord> end() {
-      finalizeOpenStreamTypeRecord();
-      return ImmutableList.copyOf(typeRecords);
-    }
-  }
-
-  /** An immutable model describing a class from a stream-based API such as RxJava. */
-  private static class StreamTypeRecord {
-
-    private final TypePredicate typePredicate;
-
-    // Names of all the methods of this type that behave like .filter(...) (must take exactly 1
-    // argument)
-    private final ImmutableSet<String> filterMethodSigs;
-    private final ImmutableSet<String> filterMethodSimpleNames;
-
-    // Names and relevant arguments of all the methods of this type that behave like .map(...) for
-    // the purposes of this checker (the listed arguments are those that take the potentially
-    // filtered objects from the stream)
-    private final ImmutableMap<String, MaplikeMethodRecord> mapMethodSigToRecord;
-    private final ImmutableMap<String, MaplikeMethodRecord> mapMethodSimpleNameToRecord;
-
-    // List of methods of java.util.stream.Stream through which we just propagate the nullability
-    // information of the last call, e.g. m() in Observable.filter(...).m().map(...) means the
-    // nullability information from filter(...) should still be propagated to map(...), ignoring the
-    // interleaving call to m().
-    // We assume that if m() is a pass-through method for Observable, so are m(a1), m(a1,a2), etc.
-    // If that is ever not the case, we can use a more complex method subsignature her.
-    private final ImmutableSet<String> passthroughMethodSigs;
-    private final ImmutableSet<String> passthroughMethodSimpleNames;
-
-    public StreamTypeRecord(
-        TypePredicate typePredicate,
-        ImmutableSet<String> filterMethodSigs,
-        ImmutableSet<String> filterMethodSimpleNames,
-        ImmutableMap<String, MaplikeMethodRecord> mapMethodSigToRecord,
-        ImmutableMap<String, MaplikeMethodRecord> mapMethodSimpleNameToRecord,
-        ImmutableSet<String> passthroughMethodSigs,
-        ImmutableSet<String> passthroughMethodSimpleNames) {
-      this.typePredicate = typePredicate;
-      this.filterMethodSigs = filterMethodSigs;
-      this.filterMethodSimpleNames = filterMethodSimpleNames;
-      this.mapMethodSigToRecord = mapMethodSigToRecord;
-      this.mapMethodSimpleNameToRecord = mapMethodSimpleNameToRecord;
-      this.passthroughMethodSigs = passthroughMethodSigs;
-      this.passthroughMethodSimpleNames = passthroughMethodSimpleNames;
-    }
-
-    public boolean matchesType(Type type, VisitorState state) {
-      return typePredicate.apply(type, state);
-    }
-
-    public boolean isFilterMethod(Symbol.MethodSymbol methodSymbol) {
-      return filterMethodSigs.contains(methodSymbol.toString())
-          || filterMethodSimpleNames.contains(methodSymbol.getQualifiedName().toString());
-    }
-
-    public boolean isMapMethod(Symbol.MethodSymbol methodSymbol) {
-      return mapMethodSigToRecord.containsKey(methodSymbol.toString())
-          || mapMethodSimpleNameToRecord.containsKey(methodSymbol.getQualifiedName().toString());
-    }
-
-    public MaplikeMethodRecord getMaplikeMethodRecord(Symbol.MethodSymbol methodSymbol) {
-      MaplikeMethodRecord record = mapMethodSigToRecord.get(methodSymbol.toString());
-      if (record == null) {
-        record = mapMethodSimpleNameToRecord.get(methodSymbol.getQualifiedName().toString());
-      }
-      return record;
-    }
-
-    public boolean isPassthroughMethod(Symbol.MethodSymbol methodSymbol) {
-      return passthroughMethodSigs.contains(methodSymbol.toString())
-          || passthroughMethodSimpleNames.contains(methodSymbol.getQualifiedName().toString());
-    }
-  }
-
-  /** An immutable model describing a map-like method from a stream-based API such as RxJava. */
-  private static class MaplikeMethodRecord {
-
-    private final String innerMethodName;
-
-    public String getInnerMethodName() {
-      return innerMethodName;
-    }
-
-    private final Set<Integer> argsFromStream;
-
-    public Set<Integer> getArgsFromStream() {
-      return argsFromStream;
-    }
-
-    public MaplikeMethodRecord(String innerMethodName, Set<Integer> argsFromStream) {
-      this.innerMethodName = innerMethodName;
-      this.argsFromStream = argsFromStream;
-    }
-  }
-
-  /**
-   * Internal bookeeping record that keeps track of the model of a map-like method and the previous
-   * filter method's inner method tree. See RxNullabilityPropagator documentation and diagram.
-   */
-  private static class MaplikeToFilterInstanceRecord {
-
-    private final MaplikeMethodRecord mapMR;
-
-    public MaplikeMethodRecord getMaplikeMethodRecord() {
-      return mapMR;
-    }
-
-    private final Tree filter;
-
-    public Tree getFilter() {
-      return filter;
-    }
-
-    public MaplikeToFilterInstanceRecord(MaplikeMethodRecord mapMR, Tree filter) {
-      assert (filter instanceof MethodTree || filter instanceof LambdaExpressionTree);
-      this.mapMR = mapMR;
-      this.filter = filter;
     }
   }
 }
